+++ http-server/inc/utils.h
@@ -1,6 +1,9 @@
 #pragma once
 #include <errno.h>
 #include <cstring>
+#include <string>
 
 #define eprintf(format, ...) \
    fprintf(stderr, format ": %s\n" __VA_OPT__(,) __VA_ARGS__, strerror(errno))
+
+std::string getIpAddrStr(struct sockaddr *addr);
